Add menu_draw.c to build
